pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "autodebias"
version = "0.1.0"
description = "自动检测和缓解文本到图像扩散模型中偏见的工具包"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "AutoDebias Team", email = "autodebias@example.com"}
]
maintainers = [
    {name = "AutoDebias Team", email = "autodebias@example.com"}
]
keywords = ["bias detection", "diffusion models", "AI fairness", "image generation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.30.0",
    "diffusers>=0.20.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "pyyaml>=6.0",
    "openai>=1.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]
gpu = [
    "accelerate>=0.20.0",
]
full = [
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",
    "tqdm>=4.65.0",
    "pandas>=2.0.0",
    "datasets>=2.14.0",
]

[project.scripts]
autodebias = "autodebias.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/autodebias"
Repository = "https://github.com/yourusername/autodebias"
Documentation = "https://autodebias.readthedocs.io/"
"Bug Reports" = "https://github.com/yourusername/autodebias/issues"

[tool.setuptools]
packages = ["autodebias", "autodebias.detectors", "autodebias.trainers", "autodebias.evaluation", "autodebias.utils"]

[tool.setuptools.package-dir]
autodebias = "."

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "gpu: marks tests that require GPU",
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true 